requirements.txt
python-dotenv
openai
google-generativeai
fastapi
uvicorn[standard]
